feat(dashboard): Add new template components and refactor imports for workflow editor
