+++ apps/dashboard/src/components/workflow-editor/steps/controls/array-field-title-template.tsx
@@ -1,0 +1,5 @@
+import { ArrayFieldTitleProps } from '@rjsf/utils';
+
+export const ArrayFieldTitleTemplate = (props: ArrayFieldTitleProps) => {
+  return <legend className="text-foreground-400 px-1 text-xs">{props.title}</legend>;
+};
